requirements.txt
streamlit==1.25.0
Pillow==9.5.0
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.1
opencv-python==4.7.0.72
lapx
